+++ src/core/search/block_max_conjunction_bulk_scorer.rs
@@ -137,8 +137,8 @@ where
 
       // First check if we have a chance of having a match based on max scores
       if has_min_comp
-        && MathUtil::sum_upper_bound(current_score + sum_other_at_1, scorers_len)
-          < self.scorable.min_competitive_score as f64
+        && (MathUtil::sum_upper_bound(current_score + sum_other_at_1, scorers_len) as f32)
+          < self.scorable.min_competitive_score
       {
         doc = self.scorers[0].iterator_mut().next_doc()?;
         continue;
@@ -170,8 +170,9 @@ where
       for (idx, iter) in other.iter_mut().enumerate() {
         // First check if we have a chance of having a match based on max scores
         if has_min_comp
-          && MathUtil::sum_upper_bound(current_score + self.sum_of_other_clauses[idx], scorers_len)
-            < self.scorable.min_competitive_score as f64
+          && (MathUtil::sum_upper_bound(current_score + self.sum_of_other_clauses[idx], scorers_len)
+            as f32)
+            < self.scorable.min_competitive_score
         {
           doc = self.scorers[0].iterator_mut().next_doc()?;
           continue 'advance_head;
